Add files via upload
We have had the data on currencies and precious metals for almost ten years. We need to predict Gold price while we have the other data(This is not forecasting of gold's price, we are simply predicting with the help of regressor models, if we need to forecast the price then we could use LSTM-long short term memory). At first, we uploaded the data to the panda's library and performed basic operations like getting the information about the data, and statistical measurements. After that, we processed the data so that we could use our model on it. Splitting the data into 80-20(train-test) and using RandomForestRegressor on the data in order to achieve the prediction model. Then we checked it for error and plotted the Actual price vs the Predicted price for the reference.